integrated react-inspector
