+++ package.json
@@ -55,6 +55,7 @@
     "mantra-core": "^1.6.1",
     "qs": "^6.2.0",
     "react-fuzzy": "^0.2.2",
+    "react-inspector": "^1.0.1",
     "react-modal": "^1.2.1",
     "redux": "^3.5.2"
   },
