+++ src/modules/ui/components/action_logger/index.js
@@ -1,11 +1,11 @@
 import React, { Component } from 'react';
-import Foldable from './foldable';
+import Inspector from 'react-inspector';
 import { baseFonts } from '../theme';
 
 const preStyle = {
   color: '#666',
   overflowY: 'auto',
-  padding: '8px',
+  padding: '8px 2px',
   boxSizing: 'border-box',
   border: '1px solid #ECECEC',
   borderRadius: 4,
@@ -34,12 +34,48 @@ const btnStyle = {
   marginLeft: 5,
 };
 
+const countStyle = {
+  display: 'inline-block',
+  backgroundColor: '#777777',
+  color: '#ffffff',
+  padding: '1px 5px',
+  borderRadius: '8px',
+};
+
+const logDivStyle = {
+  margin: 5,
+  paddingBottom: 8,
+  marginBottom:7,
+  overflow:'auto',
+  borderBottom: '1px solid #fafafa',
+};
+
+const inspectorStyle = {
+  marginLeft: 5,
+  float: 'left'
+};
+
+const countWrapper = {
+  minWidth: 20,
+  display: 'inline-block',
+  float:'left',
+  height: 19,
+  marginRight: 5,
+};
+
 class ActionLogger extends Component {
 
   getActionData() {
     return this.props.actions
       .map((action) => (
-        <Foldable key={action.id} action={action} />
+      <div style={logDivStyle} key={action.id}>
+        <div style={countWrapper}>
+          { action.count > 1 && <span style={ countStyle }>{ action.count }</span> }
+        </div>
+        <div style={inspectorStyle}>
+          <Inspector data={action.data} />
+        </div>
+      </div>
       ));
   }
